[dlflib] Reorganize directory structure
