+++ software/dlflib/include/dlflib/dlf_types.h
@@ -1,10 +1,11 @@
 #pragma once
+
 #include <Arduino.h>
 #include <FS.h>
 
 #include <vector>
 
-#include "dlf_util.h"
+#include "dlflib/utils/dlf_util.h"
 
 #define HEADER_VERSION 0
