+++ software/dlflib/include/dlflib/utils/uuid.h
@@ -5,8 +5,8 @@
 #ifndef UUID_GEN
 #define UUID_GEN
 // For a 32 bit int, returnVar must be of length 8 or greater.
-inline void IntToHex(const unsigned int inInt, char *returnVar) {
-  const char *HEXMAP = "0123456789abcdef";
+inline void IntToHex(const unsigned int inInt, char* returnVar) {
+  const char* HEXMAP = "0123456789abcdef";
   for (int i = 0; i < 8; i++) {
     // Shift each hex digit to the right, and then map it to its corresponding
     // value
@@ -16,7 +16,7 @@ inline void IntToHex(const unsigned int inInt, char *returnVar) {
 
 // returnUUID must be of length 37 or greater
 // (For the null terminator)
-inline void UUIDGen(char *returnUUID) {
+inline void UUIDGen(char* returnUUID) {
   for (int i = 0; i < 4; i++) {
     unsigned int chunk = esp_random();
     // UUID4 requires a few bits to be specific values.
