+++ software/oas-logger-v0/platformio.ini
@@ -14,8 +14,8 @@ board = adafruit_itsybitsy_esp32
 framework = arduino
 monitor_speed = 115200
 monitor_filters = esp32_exception_decoder
-lib_extra_dirs = ../dlflib
 lib_deps =
+    dlflib=symlink://../dlflib
     fastled/FastLED@^3.9.0
     mikalhart/TinyGPSPlus@^1.1.0
     fbiego/ESP32Time@^2.0.6
