+++ software/oas-logger-v1/platformio.ini
@@ -14,8 +14,8 @@ board = esp32-s3-devkitm-1
 framework = arduino
 monitor_speed = 115200
 monitor_filters = esp32_exception_decoder
-lib_extra_dirs = ../dlflib
 lib_deps =
+    dlflib=symlink://../dlflib
     fastled/FastLED@^3.9.0
     fbiego/ESP32Time@^2.0.6
     tzapu/WiFiManager@^2.0.16-rc.2
